+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### New features
 
 - Change HTTP client's user-agent for "Wizaplace-PHP-SDK/"+version
+- Added `\Wizaplace\SDK\Basket\BasketService::mergeBaskets`
 - Added `\Wizaplace\SDK\Order\OrderItem::getCustomerComment`
 
 ### Bugfixes